requirements.txt
discord.py-self
python-dotenv
keep_alive